Flatten rule processing
