+++ ayatori/src/flat_representation/ruleset.rs
@@ -1,180 +1,17 @@
 use alloc::{
     collections::{BTreeMap, BTreeSet},
     format,
-    string::{String, ToString},
+    string::String,
     vec::Vec,
 };
 use core::fmt::{self, Display};
 
-use itertools::Itertools;
-
-use super::conditions::{
-    ElementCondition, ElementConditionWithState, QuorumCondition, QuorumConditionWithState, ScalarCondition,
-    ScalarConditionWithState,
-};
+use super::rules::{Action, CollectRule, MappingRule, OnError, ScalarRule, SendBCRule, SendDMRule};
 use crate::{
-    entities::{
-        AnyTag, AnyTagRef, CollectedTag, ComputedMappingTag, ComputedScalarTag, DeserializeFunction, FullName,
-        LocalSignedBCTag, LocalSignedDMTag, MappingFunction, MappingTag, MergedScalarTag, ReceivedTag, RemoteSignedTag,
-        RuntimeError, ScalarArgumentTag, ScalarFunction, ScalarTag, SentBCTag, SentDMTag, SerdeAdapter,
-        SerializeAndSignBCFunction, SerializeAndSignDMFunction,
-    },
+    entities::{AnyTagRef, CollectedTag, ComputedScalarTag, MappingTag, RuntimeError, ScalarArgumentTag, ScalarTag},
     graph_representation::{AnyNode, GeneralizedNode, OutputNode, Reproducibility},
     traits::SessionParameters,
 };
-
-#[derive_where::derive_where(Debug)]
-struct ScalarRule<SP: SessionParameters> {
-    dependencies_condition: ScalarConditionWithState,
-    scalar_condition: ScalarConditionWithState,
-    kind: ScalarRuleKind<SP>,
-}
-
-#[derive_where::derive_where(Debug)]
-enum ScalarRuleKind<SP: SessionParameters> {
-    Compute {
-        store_in: ComputedScalarTag,
-        function: ScalarFunction<SP>,
-        args: BTreeMap<String, ScalarTag>,
-    },
-    SerializeAndSign {
-        store_in: LocalSignedBCTag,
-        function: SerializeAndSignBCFunction<SP>,
-        data: ScalarTag,
-        message_name: FullName,
-        serde_adapter: SerdeAdapter<SP::WireFormat>,
-    },
-    Merge {
-        store_in: MergedScalarTag,
-        left: ScalarTag,
-        right: ScalarTag,
-    },
-}
-
-#[derive_where::derive_where(Debug)]
-struct CollectRule<SP: SessionParameters> {
-    dependencies_condition: ScalarConditionWithState,
-    quorum_condition: QuorumConditionWithState<SP::Verifier>,
-    store_in: CollectedTag,
-    values: MappingTag,
-}
-
-#[derive_where::derive_where(Debug)]
-struct MappingRule<SP: SessionParameters> {
-    dependencies_condition: ScalarConditionWithState,
-    scalar_condition: ScalarConditionWithState,
-    element_condition: ElementConditionWithState<SP::Verifier>,
-    kind: MappingRuleKind<SP>,
-}
-
-#[derive_where::derive_where(Debug)]
-enum MappingRuleKind<SP: SessionParameters> {
-    Compute {
-        store_in: ComputedMappingTag,
-        function: MappingFunction<SP>,
-        args: BTreeMap<String, AnyTag>,
-        on_error: OnError,
-    },
-    SerializeAndSign {
-        store_in: LocalSignedDMTag,
-        function: SerializeAndSignDMFunction<SP>,
-        data: AnyTag,
-        message_name: FullName,
-        serde_adapter: SerdeAdapter<SP::WireFormat>,
-    },
-    Deserialize {
-        store_in: ReceivedTag,
-        function: DeserializeFunction<SP>,
-        data: RemoteSignedTag,
-        serde_adapter: SerdeAdapter<SP::WireFormat>,
-        expected_senders: BTreeSet<SP::Verifier>,
-        on_error: OnError,
-    },
-}
-
-#[derive_where::derive_where(Debug)]
-struct SendBCRule<SP: SessionParameters> {
-    dependencies_condition: ScalarConditionWithState,
-    scalar_condition: ScalarConditionWithState,
-    store_in: SentBCTag,
-    to_send: LocalSignedBCTag,
-    destinations: BTreeSet<SP::Verifier>,
-}
-
-#[derive_where::derive_where(Debug)]
-struct SendDMRule<SP: SessionParameters> {
-    dependencies_condition: ScalarConditionWithState,
-    element_condition: ElementConditionWithState<SP::Verifier>,
-    store_in: SentDMTag,
-    to_send: LocalSignedDMTag,
-}
-
-#[derive(Debug, Clone)]
-pub(crate) enum OnError {
-    CollectEvidence(BTreeSet<FullName>),
-    Escalate,
-}
-
-#[derive_where::derive_where(Debug)]
-pub(crate) enum Action<SP: SessionParameters> {
-    ComputeScalar {
-        store_in: ComputedScalarTag,
-        function: ScalarFunction<SP>,
-        args: BTreeMap<String, ScalarTag>,
-    },
-    ComputeMappingElement {
-        store_in: ComputedMappingTag,
-        index: SP::Verifier,
-        function: MappingFunction<SP>,
-        args: BTreeMap<String, AnyTag>,
-        on_error: OnError,
-    },
-    ComputeSerializeAndSignScalar {
-        store_in: LocalSignedBCTag,
-        function: SerializeAndSignBCFunction<SP>,
-        data: ScalarTag,
-        message_name: FullName,
-        serde_adapter: SerdeAdapter<SP::WireFormat>,
-    },
-    ComputeSerializeAndSignElement {
-        store_in: LocalSignedDMTag,
-        index: SP::Verifier,
-        function: SerializeAndSignDMFunction<SP>,
-        data: AnyTag,
-        message_name: FullName,
-        serde_adapter: SerdeAdapter<SP::WireFormat>,
-    },
-    ComputeDeserializeElement {
-        store_in: ReceivedTag,
-        index: SP::Verifier,
-        function: DeserializeFunction<SP>,
-        data: RemoteSignedTag,
-        serde_adapter: SerdeAdapter<SP::WireFormat>,
-        expected_senders: BTreeSet<SP::Verifier>,
-        on_error: OnError,
-    },
-    SendBC {
-        store_in: SentBCTag,
-        to_send: LocalSignedBCTag,
-        destinations: BTreeSet<SP::Verifier>,
-    },
-    SendDM {
-        store_in: SentDMTag,
-        to_send: LocalSignedDMTag,
-        destination: SP::Verifier,
-    },
-    Collect {
-        store_in: CollectedTag,
-        values: MappingTag,
-        // TODO: "sources"
-        indices: BTreeSet<SP::Verifier>,
-    },
-    MergeScalar {
-        store_in: MergedScalarTag,
-        left: ScalarTag,
-        right: ScalarTag,
-    },
-}
 
 fn get_on_error<SP: SessionParameters, T>(node: &T, private_inputs: &BTreeSet<String>) -> OnError
 where
@@ -302,120 +139,32 @@ impl<SP: SessionParameters> Ruleset<SP> {
         // Nodes can be iterated in any order here, but we do leaves first to make the sequence of rules more logical
         // in case someone has to look at it during debugging.
         for node in AnyNode::from(output_node.get_strong_ref()).flattened_leaves_first() {
-            let dependencies_condition =
-                ScalarConditionWithState::new(ScalarCondition::from_dependencies(node.dependencies()));
             match node {
                 AnyNode::ScalarArgument(node) => {
                     let node = node.as_ref();
                     arguments.insert(node.name.clone(), node.store_in.clone());
                 }
                 AnyNode::ComputeScalar(node) => {
-                    let node = node.as_ref();
-                    let scalar_condition = ScalarConditionWithState::new(ScalarCondition::from_compute_scalar(node));
-
-                    let function = node.function();
-
-                    let arg_tags = node
-                        .args
-                        .iter()
-                        .map(|(name, arg)| {
-                            let arg = arg.store_in().to_owned();
-                            (name.clone(), arg)
-                        })
-                        .collect();
-
-                    scalar_rules.push(ScalarRule {
-                        dependencies_condition,
-                        scalar_condition,
-                        kind: ScalarRuleKind::Compute {
-                            store_in: node.store_in.clone(),
-                            function,
-                            args: arg_tags,
-                        },
-                    });
+                    scalar_rules.push(ScalarRule::new_compute(node.as_ref()));
                 }
                 AnyNode::ComputeMapping(node) => {
                     let on_error = get_on_error(&node, private_inputs);
                     let node = node.as_ref();
                     let possible_ids = propagated_groups.get(&node.store_in)?;
-
-                    let scalar_condition = ScalarConditionWithState::new(ScalarCondition::from_compute_mapping(node));
-                    let element_condition =
-                        ElementConditionWithState::new(ElementCondition::from_compute_mapping(node), possible_ids);
-
-                    let function = node.function();
-
-                    let arg_tags = node
-                        .args
-                        .iter()
-                        .map(|(name, arg)| {
-                            let arg = arg.store_in().to_owned();
-                            (name.clone(), arg)
-                        })
-                        .collect();
-
-                    mapping_rules.push(MappingRule {
-                        dependencies_condition,
-                        scalar_condition,
-                        element_condition,
-                        kind: MappingRuleKind::Compute {
-                            store_in: node.store_in.clone(),
-                            function,
-                            args: arg_tags,
-                            on_error: on_error.clone(),
-                        },
-                    });
+                    mapping_rules.push(MappingRule::new_compute(node, possible_ids, on_error));
                 }
                 AnyNode::SerializeAndSignBC(node) => {
-                    let node = node.as_ref();
-
-                    let scalar_condition =
-                        ScalarConditionWithState::new(ScalarCondition::from_serialize_and_sign_bc(node));
-
-                    scalar_rules.push(ScalarRule {
-                        dependencies_condition,
-                        scalar_condition,
-                        kind: ScalarRuleKind::SerializeAndSign {
-                            store_in: node.store_in.clone(),
-                            function: node.function.clone(),
-                            data: node.data.store_in().to_owned(),
-                            message_name: node.message_name.clone(),
-                            serde_adapter: node.serde_adapter.clone(),
-                        },
-                    });
+                    scalar_rules.push(ScalarRule::new_serialize_and_sign(node.as_ref()));
                 }
                 AnyNode::SerializeAndSignDM(node) => {
                     let node = node.as_ref();
                     let possible_ids = propagated_groups.get(&node.store_in)?;
-
-                    let scalar_condition =
-                        ScalarConditionWithState::new(ScalarCondition::from_serialize_and_sign_dm(node));
-                    let element_condition =
-                        ElementConditionWithState::new(ElementCondition::from_serialize_and_sign(node), possible_ids);
-
-                    mapping_rules.push(MappingRule {
-                        dependencies_condition,
-                        scalar_condition,
-                        element_condition,
-                        kind: MappingRuleKind::SerializeAndSign {
-                            store_in: node.store_in.clone(),
-                            function: node.function.clone(),
-                            data: node.data.store_in().to_owned(),
-                            message_name: node.message_name.clone(),
-                            serde_adapter: node.serde_adapter.clone(),
-                        },
-                    });
+                    mapping_rules.push(MappingRule::new_serialize_and_sign(node, possible_ids));
                 }
                 AnyNode::DeserializeAndCheck(node) => {
                     let on_error = get_on_error(&node, private_inputs);
                     let node = node.as_ref();
                     let possible_ids = propagated_groups.get(&node.store_in)?;
-
-                    let scalar_condition = ScalarConditionWithState::new(ScalarCondition::empty());
-                    let element_condition = ElementConditionWithState::new(
-                        ElementCondition::from_deserialize_and_check(node),
-                        possible_ids,
-                    );
 
                     // We expect the expected senders to be present because they are added by the `Receive` node,
                     // which is an argument to this node, so it would have been processed previously.
@@ -423,66 +172,26 @@ impl<SP: SessionParameters> Ruleset<SP> {
                         RuntimeError::new(format!("Expected senders for `{}` to be available", node.message_name))
                     })?;
 
-                    mapping_rules.push(MappingRule {
-                        dependencies_condition,
-                        scalar_condition,
-                        element_condition,
-                        kind: MappingRuleKind::Deserialize {
-                            store_in: node.store_in.clone(),
-                            function: node.function.clone(),
-                            data: node.data.as_ref().store_in.clone(),
-                            serde_adapter: node.serde_adapter.clone(),
-                            expected_senders,
-                            on_error,
-                        },
-                    });
+                    mapping_rules.push(MappingRule::new_deserialize(
+                        node,
+                        expected_senders,
+                        possible_ids,
+                        on_error,
+                    ));
                 }
                 AnyNode::SendBC(node) => {
-                    let node = node.as_ref();
-
-                    let scalar_condition = ScalarConditionWithState::new(ScalarCondition::from_broadcast_message(node));
-
-                    send_bc_rules.push(SendBCRule {
-                        dependencies_condition,
-                        scalar_condition,
-                        store_in: node.store_in.clone(),
-                        to_send: node.data.as_ref().store_in.clone(),
-                        destinations: node.destinations.clone(),
-                    });
+                    send_bc_rules.push(SendBCRule::new(node.as_ref()));
                 }
                 AnyNode::SendDM(node) => {
                     let node = node.as_ref();
                     let possible_ids = propagated_groups.get(&node.store_in)?;
-
-                    let element_condition =
-                        ElementConditionWithState::new(ElementCondition::from_direct_message(node), possible_ids);
-
-                    send_dm_rules.push(SendDMRule {
-                        dependencies_condition,
-                        element_condition,
-                        store_in: node.store_in.clone(),
-                        to_send: node.data.as_ref().store_in.clone(),
-                    });
+                    send_dm_rules.push(SendDMRule::new(node, possible_ids));
                 }
                 AnyNode::Collect(node) => {
-                    let node = node.as_ref();
-                    let quorum_condition = QuorumConditionWithState::new(QuorumCondition::from_collect(node));
-                    collect_rules.push(CollectRule {
-                        dependencies_condition,
-                        quorum_condition,
-                        store_in: node.store_in.clone(),
-                        values: node.values.store_in().to_owned(),
-                    });
+                    collect_rules.push(CollectRule::new(node.as_ref()));
                 }
                 AnyNode::SendAll(node) => {
-                    let node = node.as_ref();
-                    let quorum_condition = QuorumConditionWithState::new(QuorumCondition::from_send_all(node));
-                    collect_rules.push(CollectRule {
-                        dependencies_condition,
-                        quorum_condition,
-                        store_in: node.store_in.clone(),
-                        values: MappingTag::Sent(node.values.as_ref().store_in.clone()),
-                    });
+                    collect_rules.push(CollectRule::new_send_all(node.as_ref()));
                 }
                 AnyNode::Receive(node) => {
                     let node = node.as_ref();
@@ -490,17 +199,7 @@ impl<SP: SessionParameters> Ruleset<SP> {
                     expected_messages.insert(node.message_name.clone(), possible_ids.clone());
                 }
                 AnyNode::MergeScalars(node) => {
-                    let node = node.as_ref();
-                    let scalar_condition = ScalarConditionWithState::new(ScalarCondition::from_merged_scalar(node));
-                    scalar_rules.push(ScalarRule {
-                        dependencies_condition,
-                        scalar_condition,
-                        kind: ScalarRuleKind::Merge {
-                            store_in: node.store_in.clone(),
-                            left: node.left.store_in().to_owned(),
-                            right: node.right.store_in().to_owned(),
-                        },
-                    });
+                    scalar_rules.push(ScalarRule::new_merge(node.as_ref()));
                 }
             }
         }
@@ -520,9 +219,9 @@ impl<SP: SessionParameters> Ruleset<SP> {
     pub fn update_with_banned_party(&mut self, id: &SP::Verifier) {
         let mut impossible_collects = Vec::new();
         for rule in &mut self.collect_rules {
-            rule.quorum_condition.update_with_banned_party(id);
-            if !rule.quorum_condition.is_satisfiable() {
-                impossible_collects.push(rule.store_in.clone());
+            rule.update_with_banned_party(id);
+            if !rule.is_satisfiable() {
+                impossible_collects.push(rule.store_in().clone());
             }
         }
 
@@ -539,114 +238,65 @@ impl<SP: SessionParameters> Ruleset<SP> {
         }
 
         for rule in &mut self.scalar_rules {
-            rule.dependencies_condition.update_with_scalar_ready(tag);
-            rule.scalar_condition.update_with_scalar_ready(tag);
+            rule.update_with_scalar_ready(tag);
         }
 
         for rule in &mut self.collect_rules {
-            rule.dependencies_condition.update_with_scalar_ready(tag);
+            rule.update_with_scalar_ready(tag);
         }
 
         for rule in &mut self.mapping_rules {
-            rule.dependencies_condition.update_with_scalar_ready(tag);
-            rule.scalar_condition.update_with_scalar_ready(tag);
+            rule.update_with_scalar_ready(tag);
         }
 
         for rule in &mut self.send_bc_rules {
-            rule.dependencies_condition.update_with_scalar_ready(tag);
-            rule.scalar_condition.update_with_scalar_ready(tag);
+            rule.update_with_scalar_ready(tag);
         }
 
         for rule in &mut self.send_dm_rules {
-            rule.dependencies_condition.update_with_scalar_ready(tag);
+            rule.update_with_scalar_ready(tag);
         }
     }
 
     pub fn update_with_element_ready(&mut self, tag: &MappingTag, id: &SP::Verifier) {
         for rule in &mut self.collect_rules {
-            rule.quorum_condition.update_with_element_ready(tag, id);
+            rule.update_with_element_ready(tag, id);
         }
 
         for rule in &mut self.mapping_rules {
-            rule.element_condition.update_with_element_ready(tag, id);
+            rule.update_with_element_ready(tag, id);
         }
 
         for rule in &mut self.send_dm_rules {
-            rule.element_condition.update_with_element_ready(tag, id);
+            rule.update_with_element_ready(tag, id);
         }
     }
 
     fn pop_scalar_action(&mut self) -> Option<Action<SP>> {
         self.scalar_rules
-            .extract_if(.., |rule| {
-                rule.dependencies_condition.is_satisfied() && rule.scalar_condition.is_satisfied()
-            })
+            .extract_if(.., |rule| rule.is_satisfied())
             .next()
-            .map(|rule| match rule.kind {
-                ScalarRuleKind::Compute {
-                    store_in,
-                    function,
-                    args,
-                } => Action::ComputeScalar {
-                    store_in,
-                    function,
-                    args,
-                },
-                ScalarRuleKind::SerializeAndSign {
-                    store_in,
-                    function,
-                    data,
-                    message_name,
-                    serde_adapter,
-                } => Action::ComputeSerializeAndSignScalar {
-                    store_in,
-                    function,
-                    data,
-                    message_name,
-                    serde_adapter,
-                },
-                ScalarRuleKind::Merge { store_in, left, right } => Action::MergeScalar { store_in, left, right },
-            })
+            .map(ScalarRule::into_action)
     }
 
     fn pop_collect_action(&mut self) -> Option<Action<SP>> {
         self.collect_rules
-            .extract_if(.., |rule| {
-                rule.dependencies_condition.is_satisfied() && rule.quorum_condition.is_satisfied()
-            })
+            .extract_if(.., |rule| rule.is_satisfied())
             .next()
-            .map(|rule| Action::Collect {
-                store_in: rule.store_in,
-                values: rule.values,
-                indices: rule.quorum_condition.available_ids(),
-            })
+            .map(CollectRule::into_action)
     }
 
     fn pop_bc_send_action(&mut self) -> Option<Action<SP>> {
         self.send_bc_rules
-            .extract_if(.., |rule| {
-                rule.dependencies_condition.is_satisfied() && rule.scalar_condition.is_satisfied()
-            })
+            .extract_if(.., |rule| rule.is_satisfied())
             .next()
-            .map(|rule| Action::SendBC {
-                store_in: rule.store_in.clone(),
-                to_send: rule.to_send.clone(),
-                destinations: rule.destinations,
-            })
+            .map(SendBCRule::into_action)
     }
 
     fn pop_dm_send_action(&mut self) -> Option<Action<SP>> {
         for rule in &mut self.send_dm_rules.iter_mut() {
-            if !rule.dependencies_condition.is_satisfied() {
-                continue;
-            }
-
-            if let Some(id) = rule.element_condition.pop_satisfied() {
-                return Some(Action::SendDM {
-                    store_in: rule.store_in.clone(),
-                    to_send: rule.to_send.clone(),
-                    destination: id,
-                });
+            if let Some(id) = rule.pop_satisfied() {
+                return Some(rule.make_action(id));
             }
         }
 
@@ -655,55 +305,8 @@ impl<SP: SessionParameters> Ruleset<SP> {
 
     fn pop_mapping_action(&mut self) -> Option<Action<SP>> {
         for rule in &mut self.mapping_rules.iter_mut() {
-            if !rule.dependencies_condition.is_satisfied() || !rule.scalar_condition.is_satisfied() {
-                continue;
-            }
-
-            if let Some(id) = rule.element_condition.pop_satisfied() {
-                return Some(match &rule.kind {
-                    MappingRuleKind::Compute {
-                        store_in,
-                        function,
-                        args,
-                        on_error,
-                    } => Action::ComputeMappingElement {
-                        store_in: store_in.clone(),
-                        index: id,
-                        function: function.clone(),
-                        args: args.clone(),
-                        on_error: on_error.clone(),
-                    },
-                    MappingRuleKind::SerializeAndSign {
-                        store_in,
-                        function,
-                        data,
-                        message_name,
-                        serde_adapter,
-                    } => Action::ComputeSerializeAndSignElement {
-                        store_in: store_in.clone(),
-                        index: id,
-                        function: function.clone(),
-                        data: data.clone(),
-                        message_name: message_name.clone(),
-                        serde_adapter: serde_adapter.clone(),
-                    },
-                    MappingRuleKind::Deserialize {
-                        store_in,
-                        function,
-                        data,
-                        serde_adapter,
-                        expected_senders,
-                        on_error,
-                    } => Action::ComputeDeserializeElement {
-                        store_in: store_in.clone(),
-                        index: id,
-                        function: function.clone(),
-                        data: data.clone(),
-                        serde_adapter: serde_adapter.clone(),
-                        expected_senders: expected_senders.clone(),
-                        on_error: on_error.clone(),
-                    },
-                });
+            if let Some(id) = rule.pop_satisfied() {
+                return Some(rule.make_action(id));
             }
         }
 
@@ -728,111 +331,6 @@ impl<SP: SessionParameters> Ruleset<SP> {
 
     pub fn output_tag(&self) -> &ComputedScalarTag {
         &self.output_tag
-    }
-}
-
-impl<SP: SessionParameters> Display for ScalarRule<SP> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
-        if !self.dependencies_condition.is_satisfied() {
-            writeln!(f, "if {}", self.dependencies_condition)?;
-        }
-        if !self.scalar_condition.is_satisfied() {
-            writeln!(f, "if {}", self.scalar_condition)?;
-        }
-        match &self.kind {
-            ScalarRuleKind::Compute {
-                store_in,
-                function,
-                args,
-            } => writeln!(
-                f,
-                "  {store_in} = {function}({})",
-                args.values().map(ToString::to_string).join(", ")
-            ),
-            ScalarRuleKind::SerializeAndSign {
-                store_in,
-                function,
-                data,
-                ..
-            } => writeln!(f, "{store_in} = {function}({data})"),
-            ScalarRuleKind::Merge { store_in, left, right } => writeln!(f, "  {store_in} = {left} | {right}"),
-        }
-    }
-}
-
-impl<SP: SessionParameters> Display for CollectRule<SP> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
-        if !self.dependencies_condition.is_satisfied() {
-            writeln!(f, "if {}", self.dependencies_condition)?;
-        }
-        if !self.quorum_condition.is_satisfied() {
-            writeln!(f, "if {}", self.quorum_condition)?;
-        }
-        writeln!(f, "  {} = collect({})", self.store_in, self.values)
-    }
-}
-
-impl<SP: SessionParameters> Display for MappingRule<SP> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
-        if !self.dependencies_condition.is_satisfied() {
-            writeln!(f, "if {}", self.dependencies_condition)?;
-        }
-        if !self.scalar_condition.is_satisfied() {
-            writeln!(f, "if {}", self.scalar_condition)?;
-        }
-        writeln!(f, "if {})", self.element_condition)?;
-        writeln!(f, "  {}", self.kind)
-    }
-}
-
-impl<SP: SessionParameters> Display for MappingRuleKind<SP> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
-        match self {
-            Self::Compute {
-                store_in,
-                function,
-                args,
-                ..
-            } => writeln!(
-                f,
-                "{store_in} = {function}({})",
-                args.values().map(ToString::to_string).join(", ")
-            ),
-            Self::SerializeAndSign {
-                store_in,
-                function,
-                data,
-                ..
-            } => writeln!(f, "{store_in} = {function}({data})"),
-            Self::Deserialize {
-                store_in,
-                function,
-                data,
-                ..
-            } => writeln!(f, "{store_in} = {function}({data})"),
-        }
-    }
-}
-
-impl<SP: SessionParameters> Display for SendBCRule<SP> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
-        if !self.dependencies_condition.is_satisfied() {
-            writeln!(f, "if {}", self.dependencies_condition)?;
-        }
-        if !self.scalar_condition.is_satisfied() {
-            writeln!(f, "if {}", self.scalar_condition)?;
-        }
-        writeln!(f, "  {} = broadcast_message({})", self.store_in, self.to_send)
-    }
-}
-
-impl<SP: SessionParameters> Display for SendDMRule<SP> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
-        if !self.dependencies_condition.is_satisfied() {
-            writeln!(f, "if {}", self.dependencies_condition)?;
-        }
-        writeln!(f, "if {})", self.element_condition)?;
-        writeln!(f, "  {} = direct_message({})", self.store_in, self.to_send)
     }
 }
